github_copilot_csharp_python_foundational_workshop/backend/requirements.txt
fastapi==0.109.2
uvicorn==0.27.1
gunicorn==21.2.0
sqlalchemy==2.0.27
pydantic==2.6.1
httpx==0.26.0
python-dotenv==1.0.1
python-jose[cryptography]
google-auth
requests
psycopg2-binary==2.9.9
alembic==1.13.1
